Extract common code for eval
